requirements.txt
browser-use>=0.12.0
langchain-anthropic>=0.3.0
fastapi>=0.115.0
uvicorn>=0.34.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
playwright>=1.49.0
